refactor(test): migrate agents.ts session/conversation helpers to three-tier architecture
Move 6 DB-direct seeders (ensureZeroAgentRow, createTestAgentSession,
createTestSessionWithConversation, insertTestAgentSessionWithMessages,
appendTestChatMessages, insertTestChatThread) to db-test-seeders/agents.ts
and 3 assertions (getTestSessionChatMessages,
getTestAgentSessionWithConversation, getTestAgentComposeName) to
db-test-assertions/agents.ts. Extract DB follow-up from createTestCompose
into ensureZeroAgentRow seeder. Keep 5 API/service helpers in
api-test-helpers with re-exports for backward compatibility.

Closes #9374
